pyproject.toml
[project]
name = "auth-sprint-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "async-fastapi-jwt-auth (>=0.6.6,<0.7.0)",
    "typer (>=0.20.0,<0.21.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "orjson (>=3.11.4,<4.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "user-agents (>=2.2.0,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn-worker (>=0.4.0,<0.5.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.venv'

[tool.isort]
line_length = 119
